# Backend/openapi.yml
openapi: 3.0.0
info:
  title: Beacon API
  version: 1.0.0
  description: |
    Backend API for Beacon.

    Most `/api/*` endpoints require a JWT in `Authorization: Bearer <token>`.
    Request validation is enforced via `express-openapi-validator`.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/pins/user:
    get:
      tags: [Pins]
      summary: Get all pins created by the logged-in user
      description: Retrieves all pins created by the authenticated user
      operationId: getUserPins
      responses:
        200:
          description: Successful response with array of user's pins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pin'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/me:
    patch:
      tags: [Auth]
      summary: Update current user profile
      description: Update bio, avatar URL, or profile visibility
      operationId: updateMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bio:
                  type: string
                  maxLength: 300
                  nullable: true
                avatar:
                  type: string
                  maxLength: 2000
                  nullable: true
                profileVisibility:
                  type: string
                  enum: [public, friends, private]
      responses:
        200:
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/users/{userID}:
    get:
      tags: [Auth]
      summary: Get user profile
      description: Returns a user's public profile. Private profiles are hidden from non-followers.
      operationId: getUser
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        403:
          description: Profile is private
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/users/{userID}/followers:
    get:
      tags: [Auth]
      summary: Get user's followers
      description: Returns paginated list of followers for a user
      operationId: getUserFollowers
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        200:
          description: Paginated followers list
          content:
            application/json:
              schema:
                type: object
                properties:
                  followers:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserSummary'
                  page: { type: integer }
                  hasMore: { type: boolean }
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/users/{userID}/following:
    get:
      tags: [Auth]
      summary: Get users that a user follows
      description: Returns paginated list of users that the given user follows
      operationId: getUserFollowing
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        200:
          description: Paginated following list
          content:
            application/json:
              schema:
                type: object
                properties:
                  following:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserSummary'
                  page: { type: integer }
                  hasMore: { type: boolean }
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pin:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the pin
          example: 1
        creatorID:
          type: integer
          description: ID of the account that created the pin
          example: 1
        email:
          type: string
          format: email
          description: Email of the pin creator
          example: "user@example.com"
        latitude:
          type: number
          description: self explanatory
        longitude:
          type: number
          description: self explanatory
        title:
          type: string
          maxLength: 200
          description: Title of the pin
          example: "Golden Gate Overlook"
          nullable: true
        description:
          type: string
          maxLength: 500
          description: Description/message content of the pin
          example: "Welcome to Beacon!"
          nullable: true
        image:
          type: string
          maxLength: 2000
          description: URL of the image
          example: "https://images.unsplash.com/photo-1516192518150-0d8fee5425e3"
          nullable: true
        address:
          type: string
          maxLength: 200
          description: Address of the pin location
          example: "123 Main St, San Francisco, CA"
          nullable: true
        tags:
          type: string
          maxLength: 500
          description: Tags stored as a JSON-encoded string array
          example: '["scenic","viewpoint","nature"]'
          nullable: true
        likes:
          type: integer
          description: Base like count stored on the pin
          example: 0
          nullable: true
        createdAt:
          type: string
          description: Timestamp when the pin was created
          example: "2026-01-17 12:00:00"
          nullable: true
        userStatus:
          type: string
          enum: [visited, wishlist]
          nullable: true
          description: |
            The requesting user's status for this pin, if any. Only populated on
            authenticated endpoints.
        trendingScore:
          type: number
          nullable: true
          description: |
            Computed trending score `likes + 3 * max(0, 1 - ageDays / days)`.
            Only present on `/api/pins/trending`.
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Human-readable error message
          example: "Invalid token"
        error:
          oneOf:
            - type: string
            - type: object
          description: Error identifier/message (some endpoints return 'error' instead of 'message')
          example: "Failed to share itinerary"
        errors:
          type: array
          description: Detailed validation errors (OpenAPI validator)
          items:
            type: object
        details:
          type: string
          description: Additional error details
          example: "FOREIGN KEY constraint failed"
...
    UserProfile:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        bio: { type: string, nullable: true }
        avatar: { type: string, nullable: true }
        profileVisibility: { type: string, enum: [public, friends, private] }
        followerCount: { type: integer }
        followingCount: { type: integer }
        isFollowed: { type: boolean }
        isOwn: { type: boolean }
    UserSummary:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        bio: { type: string, nullable: true }
        avatar: { type: string, nullable: true }
        followerCount: { type: integer }
        isFollowed: { type: boolean }
...